requirement_airflow.txt
websocket-client
confluent_kafka
pyhive
pyspark
apache-airflow-providers-apache-spark
thrift
thrift_sasl